#7 #8 feature, refactor: chat 테이블에 colunm 추가, chatmessage 패키지 이름 변경
