+++ src/main/java/com/chatting/room/chat/api/ChatMessageController.java
@@ -1,0 +1,4 @@
+package com.chatting.room.chat.api;
+
+public class ChatMessageController {
+}
