+++ src/main/java/com/chatting/room/chat/application/ChatMessageService.java
@@ -1,0 +1,4 @@
+package com.chatting.room.chat.application;
+
+public class ChatMessageService {
+}
