+++ src/main/java/com/chatting/room/chat/domain/ChatMessage.java
@@ -1,0 +1,4 @@
+package com.chatting.room.chat.domain;
+
+public class ChatMessage {
+}
