+++ src/main/java/com/chatting/room/chat/dto/request/chatMessageReq.java
@@ -1,0 +1,4 @@
+package com.chatting.room.chat.dto.request;
+
+public class chatMessageReq {
+}
