+++ src/main/java/com/chatting/room/chat/dto/response/ChatMessageResp.java
@@ -1,0 +1,4 @@
+package com.chatting.room.chat.dto.response;
+
+public class ChatMessageResp {
+}
