+++ src/main/java/com/chatting/room/chat/execption/ChatMessageException.java
@@ -1,0 +1,4 @@
+package com.chatting.room.chat.execption;
+
+public class ChatMessageException {
+}
