+++ src/main/java/com/chatting/room/chat/repository/ChatMessageRepository.java
@@ -1,0 +1,4 @@
+package com.chatting.room.chat.repository;
+
+public class ChatMessageRepository {
+}
