--- src/main/java/com/chatting/room/chatmessage/api/ChatMessageController.java
@@ -1,4 +1,0 @@
-package com.chatting.room.chatmessage.api;
-
-public class ChatMessageController {
-}
