--- src/main/java/com/chatting/room/chatmessage/application/ChatMessageService.java
@@ -1,4 +1,0 @@
-package com.chatting.room.chatmessage.application;
-
-public class ChatMessageService {
-}
