--- src/main/java/com/chatting/room/chatmessage/domain/ChatMessage.java
@@ -1,4 +1,0 @@
-package com.chatting.room.chatmessage.domain;
-
-public class ChatMessage {
-}
