--- src/main/java/com/chatting/room/chatmessage/dto/request/chatMessageReq.java
@@ -1,4 +1,0 @@
-package com.chatting.room.chatmessage.dto.request;
-
-public class chatMessageReq {
-}
