--- src/main/java/com/chatting/room/chatmessage/dto/response/ChatMessageResp.java
@@ -1,4 +1,0 @@
-package com.chatting.room.chatmessage.dto.response;
-
-public class ChatMessageResp {
-}
