--- src/main/java/com/chatting/room/chatmessage/execption/ChatMessageException.java
@@ -1,4 +1,0 @@
-package com.chatting.room.chatmessage.execption;
-
-public class ChatMessageException {
-}
