--- src/main/java/com/chatting/room/chatmessage/repository/ChatMessageRepository.java
@@ -1,4 +1,0 @@
-package com.chatting.room.chatmessage.repository;
-
-public class ChatMessageRepository {
-}
