+++ src/main/resources/db/V2_add_password_col_to_chatroon.sql
@@ -1,0 +1,1 @@
+alter table chatroom add password varchar(4) after 'description';
